Derive Debug for Additional Types
With gfx-hal 0.2.1, all of the gfx-hal types that wgpu depends on
implement Debug. Thus, some types that could not derive Debug in #216
can now derive Debug.

The sole public type that does not derive Debug at this point is
wgpu-native::hub::Hub. It depends on backend types that do not yet
implement Debug.

Related to #76.